Add selecting Float/Double Tensor;GAE now only Float Tensor

+++ gae/utils.py
@@ -49,7 +49,8 @@ def load_data(dataset):
 
     features = sp.vstack((allx, tx)).tolil()
     features[test_idx_reorder, :] = features[test_idx_range, :]
-    features = torch.DoubleTensor(np.array(features.todense()))
+    # features = torch.DoubleTensor(np.array(features.todense()))
+    features = torch.FloatTensor(np.array(features.todense()))
     adj = nx.adjacency_matrix(nx.from_dict_of_lists(graph))
 
     labels = np.vstack((ally, ty))
@@ -273,12 +274,14 @@ def preprocess_graph(adj):
 
 def sparse_mx_to_torch_sparse_tensor(sparse_mx):
     """Convert a scipy sparse matrix to a torch sparse tensor."""
-    sparse_mx = sparse_mx.tocoo().astype(np.float64)
+    # sparse_mx = sparse_mx.tocoo().astype(np.float64)
+    sparse_mx = sparse_mx.tocoo().astype(np.float32)
     indices = torch.from_numpy(
         np.vstack((sparse_mx.row, sparse_mx.col)).astype(np.int64))
     values = torch.from_numpy(sparse_mx.data)
     shape = torch.Size(sparse_mx.shape)
-    return torch.sparse.DoubleTensor(indices, values, shape)
+    # return torch.sparse.DoubleTensor(indices, values, shape)
+    return torch.sparse.FloatTensor(indices, values, shape)
 
 
 def get_roc_score(emb, adj_orig, edges_pos, edges_neg):
